pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "biopython>=1.85",
    "google-genai>=1.46.0",
    "openai>=2.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "spyprot>=0.9.6",
    "streamlit>=1.50.0",
]
[server]
port = 8501